+++ lib/zorb/interpreter.ex
@@ -2530,6 +2530,28 @@ defmodule Zorb.Interpreter do
     halt(reason, @pc, 0)
   end
 
+  # --- Orbit Radio Metadata ---
+
+  defw zio_get_version(), I32 do
+    @version
+  end
+
+  defw zio_get_serial(), T.Address do
+    I32.const(0x83001)
+  end
+
+  defw zio_get_command_prefix(), I32 do
+    read_byte(0x83008)
+  end
+
+  defw zio_get_chat_prefix(), T.Address do
+    I32.const(0x83010)
+  end
+
+  defw zio_get_channel_prefix(), T.Address do
+    I32.const(0x83050)
+  end
+
   defw handle_interrupts(), i: I32 do
     i = Host.check_interrupt()